Match APN types with ignoring the case
Currently APN types which are set on ApnEditor are compared with the
types which are configured with read_only_apn_types_string_array in a
case-sensitive manner. So upper case types such as "DUN" or "IMS" can be
set if read_only_apn_types_string_array is configured as "dun" or "ims."
Then the APN with types such as "DUN" or "IMS" can be added and may be
used for the network connection unintentionally.

This patch compares APN types with ignoring the case on ApnEditor to
prevent the problem above.

Bug: 200194310
Test: Manual test passed
Google: 2671056
Change-Id: I0f68bf470699df388855ec7277c0cfc24a2c30ba
Signed-off-by: Edwiin Kusuma Jaya <kutemeikito0905@gmail.com>